python/requirements.txt
redis>=5.0.0
requests>=2.31.0
gTTS>=2.5.0
openai-whisper>=20231117
opencv-python-headless>=4.9.0
imagehash>=4.3.1
Pillow>=10.0.0
librosa>=0.10.0
soundfile>=0.12.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
aiofiles>=23.0.0
